Time: 1636 ms (13.05%), Space: 259.8 MB (29.44%) - LeetHub

+++ 0446-arithmetic-slices-ii-subsequence/0446-arithmetic-slices-ii-subsequence.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int numberOfArithmeticSlices(vector<int>& nums) {
-        vector<map<long long,long long>> dp(nums.size());
+        vector<map<long long,int>> dp(nums.size());
         long long sum = 0;
         long long n = nums.size();
         for(int i = 1; i < n; i++)
